backend/requirements.txt
langchain
langchain-openai
langchain-google-genai
python-dotenv
fastapi
uvicorn
chromadb
langchain-chroma
